修改kcp_input两个参数，因为x86跟x86_64的long型长度不同，导致其它语言封送参数麻烦，所以统一使用int，长度为32bit

+++ jni/kcp.c
@@ -746,7 +746,7 @@ void ikcp_parse_data(ikcpcb *kcp, IKCPSEG *newseg)
 //---------------------------------------------------------------------
 // input data
 //---------------------------------------------------------------------
-int ikcp_input(ikcpcb *kcp, const char *data, long offset, long size)
+int ikcp_input(ikcpcb *kcp, const char *data, int offset, int size)
 {
 	IUINT32 una = kcp->snd_una;
 	IUINT32 maxack = 0;
